+++ src/Domain/Client/SubscriptionStorageInterface.php
@@ -8,9 +8,11 @@ use Closure;
 
 interface SubscriptionStorageInterface
 {
-    public function add(string $sid, Closure $closure): void;
+    public function add(string $sid, Closure $closure, string|null $queueGroup = null): void;
 
     public function get(string $sid): Closure|null;
+
+    public function getQueueGroup(string $sid): string|null;
 
     public function remove(string $sid): void;